feat: Add create method to ProfessionalService
This commit adds a `create` method to the `ProfessionalService` trait in the `professional/ports.rs` file. The `create` method returns a future that resolves to a result of `Professional` or `ProfessionalError`. This method allows the creation of a professional.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature, fix for bug fix, chore for maintenance tasks) and provides a clear and concise description of the changes made.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>