chore: Update dependencies in Cargo.toml
This commit updates the dependencies in the Cargo.toml file. It adds the "reqwest" and "tower" dependencies to the project. These dependencies are required for implementing new features related to networking and HTTP requests.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature, fix for bug fix, chore for maintenance tasks) and provides a clear and concise description of the changes made.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>