feat: Add Keycloak authentication layer to HTTP server
This commit adds the Keycloak authentication layer to the HTTP server. It introduces the `KeycloakAuthInstance` and `KeycloakAuthLayer` from the `axum_keycloak_auth` crate. The `KeycloakAuthInstance` is initialized with the Keycloak server URL and realm, and the `KeycloakAuthLayer` is added to the server's router. This layer enables authentication and authorization for the API routes.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature, fix for bug fix, chore for maintenance tasks) and provides a clear and concise description of the changes made.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>