Implement Keycloak Layer (#3)
* chore: Update dependencies in Cargo.toml

This commit updates the dependencies in the Cargo.toml file. It adds the "reqwest" and "tower" dependencies to the project. These dependencies are required for implementing new features related to networking and HTTP requests.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature, fix for bug fix, chore for maintenance tasks) and provides a clear and concise description of the changes made.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add Keycloak authentication layer to HTTP server

This commit adds the Keycloak authentication layer to the HTTP server. It introduces the `KeycloakAuthInstance` and `KeycloakAuthLayer` from the `axum_keycloak_auth` crate. The `KeycloakAuthInstance` is initialized with the Keycloak server URL and realm, and the `KeycloakAuthLayer` is added to the server's router. This layer enables authentication and authorization for the API routes.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature, fix for bug fix, chore for maintenance tasks) and provides a clear and concise description of the changes made.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Update server main.rs to use environment variables for keycloak

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

---------

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>