feat: Remove unused `layers` module from `http.rs`
The `layers` module in `http.rs` is no longer used and has been removed. This commit cleans up the codebase by removing the unused module.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature, fix for bug fix, chore for maintenance tasks) and provides a clear and concise description of the changes made.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>