feat: Add shared module for user identity in domain
This commit adds a new shared module in the domain package to handle user identity. It introduces the `UserIdentityInfo` struct to represent user identity information, and the `UserIdentityError` enum to handle errors related to user identity. Additionally, it defines the `UserIdentityProvider` trait to provide functionality for creating and deleting users. These changes lay the foundation for managing user identity in the application.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>