feat: Add Keycloak authentication module
This commit adds the Keycloak authentication module to the infrastructure package. It introduces the `KeycloakAdapter` struct, which provides functionality for interacting with the Keycloak server. The module includes methods for retrieving an access token and creating a user in Keycloak. These changes lay the foundation for integrating Keycloak authentication into the application.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>